+++ lib/libsesame3bt-core/os2.h
@@ -17,22 +17,24 @@ class OS2Handler {
 	    : client(client), transport(transport), crypt(crypt) {}
 	OS2Handler(const OS2Handler&) = delete;
 	OS2Handler& operator=(const OS2Handler&) = delete;
-	bool init() { return Ecc::initialized(); }
-	bool set_keys(std::string_view pk_str, std::string_view secret_str);
-	bool set_keys(const std::array<std::byte, Sesame::PK_SIZE>& public_key,
-	              const std::array<std::byte, Sesame::SECRET_SIZE>& secret_key);
-	bool send_command(Sesame::op_code_t op_code,
-	                  Sesame::item_code_t item_code,
-	                  const std::byte* data,
-	                  size_t data_size,
-	                  bool is_crypted);
+	result_t init() { return Ecc::initialized() ? result_t::success : result_t::crypt_failure; }
+	result_t set_keys(std::string_view pk_str, std::string_view secret_str);
+	result_t set_keys(const std::array<std::byte, Sesame::PK_SIZE>& public_key,
+	                  const std::array<std::byte, Sesame::SECRET_SIZE>& secret_key);
+	result_t send_command(Sesame::op_code_t op_code,
+	                      Sesame::item_code_t item_code,
+	                      const std::byte* data,
+	                      size_t data_size,
+	                      bool is_crypted);
 
-	void handle_publish_initial(const std::byte* in, size_t in_len);
-	void handle_response_login(const std::byte* in, size_t in_len);
-	void handle_publish_mecha_setting(const std::byte* in, size_t in_len);
-	void handle_publish_mecha_status(const std::byte* in, size_t in_len);
-	void handle_response_mecha_status(const std::byte* in, size_t in_len) { handle_publish_mecha_status(in + 2, in_len - 2); };
-	void handle_history(const std::byte* in, size_t in_len);
+	result_t handle_publish_initial(const std::byte* in, size_t in_len);
+	result_t handle_response_login(const std::byte* in, size_t in_len);
+	result_t handle_publish_mecha_setting(const std::byte* in, size_t in_len);
+	result_t handle_publish_mecha_status(const std::byte* in, size_t in_len);
+	result_t handle_response_mecha_status(const std::byte* in, size_t in_len) {
+		return handle_publish_mecha_status(in + 2, in_len - 2);
+	};
+	result_t handle_history(const std::byte* in, size_t in_len);
 	size_t get_max_history_tag_size() const { return MAX_HISTORY_TAG_SIZE; }
 	size_t get_cmd_tag_size(size_t tag_len) const { return MAX_HISTORY_TAG_SIZE + 1; }
 	static constexpr size_t MAX_HISTORY_TAG_SIZE = 21;