+++ lib/libsesame3bt-core/transport.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <cstddef>
+#include <optional>
 #include "crypt.h"
 #include "libsesame3bt/BLEBackend.h"
 
@@ -24,30 +25,28 @@ class SesameBLEBuffer {
 	SesameBLEBuffer() { reset(); }
 	void reset() {
 		recv_size = 0;
-		skipping = false;
+		prev_result.reset();
 	}
 
  private:
 	std::array<std::byte, MAX_RECV> recv_buffer{};
 	size_t recv_size;
-	bool skipping;
+	std::optional<result_t> prev_result;
 };
 
 class SesameBLETransport {
  public:
-	enum class decode_result_t { skipping, received, require_more, dropped };
 	SesameBLETransport(SesameBLEBackend& backend) : backend(backend) {}
 	SesameBLETransport(const SesameBLETransport&) = delete;
 	SesameBLETransport& operator=(const SesameBLETransport&) = delete;
 	bool send_data(const std::byte* pkt, size_t pkt_size, bool is_crypted);
-	bool send_notify(Sesame::op_code_t op_code,
-	                 Sesame::item_code_t item_code,
-	                 const std::byte* data,
-	                 size_t data_size,
-	                 bool is_crypted,
-	                 CryptHandler& crypt);
-	decode_result_t decode(const std::byte* data, size_t size, CryptHandler& crypt);
-	void disconnect();
+	result_t send_notify(Sesame::op_code_t op_code,
+	                     Sesame::item_code_t item_code,
+	                     const std::byte* data,
+	                     size_t data_size,
+	                     bool is_crypted,
+	                     CryptHandler& crypt);
+	std::optional<result_t> decode(const std::byte* data, size_t size, CryptHandler& crypt);
 	void reset();
 	const std::byte* data() { return buffer.recv_buffer.data(); }
 	size_t data_size() { return buffer.recv_size; }